Update packages + content update to be more wc3 friendly
This upgrades all packages, including font awesome to version 5.

This adds Battle.net to the WC3 setup guide as a viable place for
warlock games.

Also many small improvements and fixes.

Still not done: Update the front page about text to be a bit more
reflective on the current development status.