Remove the final paragraph from the front page
But keep it in the about page, reintroducing the "Read more" button.

Including that paragraph made the text on the front page end on too
gloomy a note to then bombard the user with those big social buttons.
This is a little better, probably, since it's still useful information
to have somewhere.