Webpack: Attempt to optimize building
- Always use import instead of commonjs require where applicable
- Load fontawesome icons once on load and pass them to templates,
  getting rid of any dynamic module require (don't do this!)
- Disable non-error messages for dev server

The webpack dev server still eats up memory fast but doesn't seem to be
doing anything that causes it to go over (when nodejs limit is reached
it trashes memory it doesn't seem to need, but it still reaches it
fairly quickly). I think it needs around ~200MB minimum right now.